internal/worker: don't call pkgsite for known non-modules
Requests to pkg.go.dev are throttled to qps, so they are expensive.
Many of the calls are unnecessary because the paths are known not
to be Go modules.

Keep a list of patterns that match such paths and check them before
calling pkgsite.

Change-Id: I5e80d0c494510824939964b72466605ecc5cacbe
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/370394
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>