cmd/govulncheck: add a Source function to internal/govulncheck
Move the call to vulncheck.Source and the subsequent removal of
uncalled vulns to a separate function, for sharing with gopls.

This fixes what I believe was a bug. The JSON output did not match the
other output modes, because it output all the vulns, not just the ones
that were called. We want the JSON to produce the same information as
other modes, just in JSON format.

Change-Id: I3b3d1b31044530079e03969374462384d069d71c
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/406935
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>