vulncheck/internal/gosym: use FuncSymName based on Go version in binary
We were choosing FuncSymName based on the Go version used to compile
govulncheck, instead of the version used to build the binary. This can
cause errors when the govulncheck and the binary to be analyzed are
built using different versions.

Change-Id: I1068919d68d4d433dfe9268b60bc598c6fb4db86
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/426934
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>