exp/govulncheck: add part of experimental govulncheck API
Gopls wants to invoke the govulncheck command line tool to get
high-level summary of vulnerability scanning. Then it will translate
any findings to LSP messages.

The govulncheck command line tool is under active development
and there is no stable API built around it yet. While govulncheck
is evolving, it can break the assumption a released version of gopls
made any time. When users independently install gopls and govulncheck,
it is hard to keep them compatible. There could be many different ways
of solving this problem, but we think it is the easiest to embed the
govulncheck logic in the gopls. This is basically equivalent to
pin the version of govulncheck. We will evaluate different approaches
(e.g. invoke govulncheck found from PATH and hope it works) as the
govulncheck command line tool interface becomes stable.

Main is a wrapper of govulncheck command's main. This never returns.

Change-Id: I050cf114827bde3f3450e06909d1501f381804c3
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/435902
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>