cmd/govulncheck: use text/template for printing to screen
This replaces the current, more error prone, approach of printing
things to screen using fmt functions.

Functions such as indent and wrap are passed to the template and used
within the template. This way, the functionalities for determining what
is printed to screen are kept closer to each other.

This CL will be followed by CLs that refactor the unit tests.

Change-Id: I296d30959f1fc117a95f40183287b5df0f0f5a82
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/459795
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>