cmd/govulncheck: fix bug in platforms row when printing results
When there are specific OSs for a vuln but no architecure, and vice
versa, we would not printing anything in the platforms row. This CL
fixes that issue.

In such cases, we just list the OSs, or archs in vice versa case. An
alternative was to print something like "windows/*", but that seemed
confusing. Having a row like "Platforms: windows" seams clear to the
user.

Change-Id: I41b9865353571b8d225d198ba172a566fa992852
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/460423
Reviewed-by: Julie Qiu <julieqiu@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>