all: remove use of golang.org/x/exp
We shouldn't depend on packages in golang.org/x/exp in x/vuln, since
code in that repository is intended to be experimental.

Copy the functions we use in x/vuln, and remove x/exp as a dependency.

Also add a check for this in all_test.go.

Change-Id: I1d5ebd1958580121282ab04790d53ccd1e20eebf
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/471079
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Julie Qiu <julieqiu@google.com>